+++ lib/src/api/authentication/session.dart
@@ -1,4 +1,7 @@
 abstract class Session {
+  /// Timestamp of the start of this session.
   DateTime get timestamp;
+
+  /// Emits an event as soon as this session expires.
   Stream<void> get expiration;
 }